mhcrnn - with errors
