+++ lightning_pose/models/heads/__init__.py
@@ -1,5 +1,5 @@
 from lightning_pose.models.heads.heatmap import HeatmapHead
-# from lightning_pose.models.heads.heatmap_mhcrnn import HeatmapMHCRNNHead
+from lightning_pose.models.heads.heatmap_mhcrnn import HeatmapMHCRNNHead
 from lightning_pose.models.heads.regression import LinearRegressionHead
 
 # to ignore imports for sphix-autoapidoc
