+++ lightning_pose/models/heatmap_tracker_mhcrnn.py
@@ -3,9 +3,7 @@
 from typing import Any, Tuple, Union
 
 import torch
-from kornia.geometry.subpix import spatial_softmax2d
 from omegaconf import DictConfig
-from torch import nn
 from torchtyping import TensorType
 from typeguard import typechecked
 from typing_extensions import Literal
@@ -24,12 +22,12 @@ from lightning_pose.models.base import (
     SemiSupervisedTrackerMixin,
     convert_bbox_coords,
 )
+from lightning_pose.models.heads import HeatmapMHCRNNHead
 
 # to ignore imports for sphix-autoapidoc
 __all__ = [
     "HeatmapTrackerMHCRNN",
     "SemiSupervisedHeatmapTrackerMHCRNN",
-    "UpsamplingCRNN",
 ]
 
 
@@ -91,30 +89,10 @@ class HeatmapTrackerMHCRNN(HeatmapTracker):
             **kwargs,
         )
 
-        # create upsampling layers for crnn
-        self.crnn = UpsamplingCRNN(
-            num_filters_for_upsampling=self.num_filters_for_upsampling,
-            num_keypoints=self.num_keypoints,
+        self.multihead = HeatmapMHCRNNHead(
+            single_frame_head=self.head,
             upsampling_factor=1 if "vit" in backbone else 2,
         )
-        self.upsampling_layers_rnn = self.crnn.layers
-
-        # alias parent upsampling layers for single frame
-        self.upsampling_layers_sf = self.upsampling_layers
-
-    def heatmaps_from_representations(
-        self,
-        representations: TensorType["batch", "features", "rep_height", "rep_width", "frames"],
-    ) -> Tuple[
-            TensorType["batch", "num_keypoints", "heatmap_height", "heatmap_width"],
-            TensorType["batch", "num_keypoints", "heatmap_height", "heatmap_width"],
-    ]:
-        """Handle context frames then upsample to get final heatmaps."""
-        # permute to shape (frames, batch, features, rep_height, rep_width)
-        representations = torch.permute(representations, (4, 0, 1, 2, 3))
-        heatmaps_crnn = self.crnn(representations)
-        heatmaps_sf = self.upsampling_layers_sf(representations[2])  # index 2 == middle frame
-        return heatmaps_crnn, heatmaps_sf
 
     def forward(
         self,
@@ -159,25 +137,13 @@ class HeatmapTrackerMHCRNN(HeatmapTracker):
             shape_r = representations.shape
             num_frames -= 4  # we lose the first/last 2 frames of unlabeled batch due to context
             representations = representations.reshape(
-                num_frames * shape[1],  -1, shape_r[-3], shape_r[-2], shape_r[-1],
+                num_frames * shape[1], -1, shape_r[-3], shape_r[-2], shape_r[-1],
             )
 
         # get two heatmaps for each representation (context, non-context)
-        heatmaps_crnn, heatmaps_sf = self.heatmaps_from_representations(representations)
+        heatmaps_crnn, heatmaps_sf = self.multihead(representations, shape, num_frames)
 
-        if len(shape) == 6 or len(shape) == 5:
-            # reshape the outputs to extract the view dimension
-            heatmaps_crnn = heatmaps_crnn.reshape(
-                num_frames, -1, heatmaps_crnn.shape[-2], heatmaps_crnn.shape[-1])
-            heatmaps_sf = heatmaps_sf.reshape(
-                num_frames, -1, heatmaps_sf.shape[-2], heatmaps_sf.shape[-1])
-
-        # normalize heatmaps
-        # softmax temp stays 1 here; to modify for model predictions, see constructor
-        heatmaps_crnn_norm = spatial_softmax2d(heatmaps_crnn, temperature=torch.tensor([1.0]))
-        heatmaps_sf_norm = spatial_softmax2d(heatmaps_sf, temperature=torch.tensor([1.0]))
-
-        return heatmaps_crnn_norm, heatmaps_sf_norm
+        return heatmaps_crnn, heatmaps_sf
 
     def get_loss_inputs_labeled(
         self,
@@ -190,8 +156,8 @@ class HeatmapTrackerMHCRNN(HeatmapTracker):
         # images -> heatmaps
         pred_heatmaps_crnn, pred_heatmaps_sf = self.forward(batch_dict["images"])
         # heatmaps -> keypoints
-        pred_keypoints_crnn, confidence_crnn = self.run_subpixelmaxima(pred_heatmaps_crnn)
-        pred_keypoints_sf, confidence_sf = self.run_subpixelmaxima(pred_heatmaps_sf)
+        pred_keypoints_crnn, confidence_crnn = self.head.run_subpixelmaxima(pred_heatmaps_crnn)
+        pred_keypoints_sf, confidence_sf = self.head.run_subpixelmaxima(pred_heatmaps_sf)
         return {
             "heatmaps_targ": torch.cat([batch_dict["heatmaps"], batch_dict["heatmaps"]], dim=0),
             "heatmaps_pred": torch.cat([pred_heatmaps_crnn, pred_heatmaps_sf], dim=0),
@@ -230,8 +196,8 @@ class HeatmapTrackerMHCRNN(HeatmapTracker):
         # images -> heatmaps
         pred_heatmaps_crnn, pred_heatmaps_sf = self.forward(images)
         # heatmaps -> keypoints
-        pred_keypoints_crnn, confidence_crnn = self.run_subpixelmaxima(pred_heatmaps_crnn)
-        pred_keypoints_sf, confidence_sf = self.run_subpixelmaxima(pred_heatmaps_sf)
+        pred_keypoints_crnn, confidence_crnn = self.head.run_subpixelmaxima(pred_heatmaps_crnn)
+        pred_keypoints_sf, confidence_sf = self.head.run_subpixelmaxima(pred_heatmaps_sf)
         # reshape keypoints to be (batch, n_keypoints, 2)
         pred_keypoints_sf = pred_keypoints_sf.reshape(pred_keypoints_sf.shape[0], -1, 2)
         pred_keypoints_crnn = pred_keypoints_crnn.reshape(pred_keypoints_crnn.shape[0], -1, 2)
@@ -254,10 +220,10 @@ class HeatmapTrackerMHCRNN(HeatmapTracker):
         params = [
             {"params": self.backbone.parameters(), "name": "backbone", "lr": 0.0},
             {
-                "params": self.upsampling_layers_rnn.parameters(),
+                "params": self.multihead.head_mf.layers.parameters(),
                 "name": "upsampling_rnn",
             },
-            {"params": self.upsampling_layers_sf.parameters(), "name": "upsampling_sf"},
+            {"params": self.multihead.head_sf.parameters(), "name": "upsampling_sf"},
         ]
         return params
 
@@ -274,7 +240,6 @@ class SemiSupervisedHeatmapTrackerMHCRNN(SemiSupervisedTrackerMixin, HeatmapTrac
         backbone: ALLOWED_BACKBONES = "resnet50",
         downsample_factor: Literal[2, 3] = 2,
         pretrained: bool = True,
-        output_shape: tuple | None = None,
         torch_seed: int = 123,
         optimizer: str = "Adam",
         optimizer_params: DictConfig | dict | None = None,
@@ -294,8 +259,6 @@ class SemiSupervisedHeatmapTrackerMHCRNN(SemiSupervisedTrackerMixin, HeatmapTrac
                 save memory; subpixel operations are performed for increased
                 precision
             pretrained: True to load pretrained imagenet weights
-            output_shape: hard-coded image size to avoid dynamic shape
-                computations
             torch_seed: make weight initialization reproducible
             lr_scheduler: how to schedule learning rate
                 multisteplr
@@ -309,7 +272,6 @@ class SemiSupervisedHeatmapTrackerMHCRNN(SemiSupervisedTrackerMixin, HeatmapTrac
             backbone=backbone,
             downsample_factor=downsample_factor,
             pretrained=pretrained,
-            output_shape=output_shape,
             torch_seed=torch_seed,
             optimizer=optimizer,
             optimizer_params=optimizer_params,
@@ -338,8 +300,8 @@ class SemiSupervisedHeatmapTrackerMHCRNN(SemiSupervisedTrackerMixin, HeatmapTrac
         )
 
         # heatmaps -> keypoints
-        pred_keypoints_crnn, confidence_crnn = self.run_subpixelmaxima(pred_heatmaps_crnn)
-        pred_keypoints_sf, confidence_sf = self.run_subpixelmaxima(pred_heatmaps_sf)
+        pred_keypoints_crnn, confidence_crnn = self.head.run_subpixelmaxima(pred_heatmaps_crnn)
+        pred_keypoints_sf, confidence_sf = self.head.run_subpixelmaxima(pred_heatmaps_sf)
 
         # undo augmentations if needed
         pred_keypoints_crnn = undo_affine_transform_batch(
@@ -362,155 +324,3 @@ class SemiSupervisedHeatmapTrackerMHCRNN(SemiSupervisedTrackerMixin, HeatmapTrac
             "keypoints_pred": torch.cat([pred_keypoints_crnn, pred_keypoints_sf], dim=0),
             "confidences": torch.cat([confidence_crnn, confidence_sf], dim=0),
         }
-
-
-class UpsamplingCRNN(torch.nn.Module):
-    """Bidirectional Convolutional RNN network that handles heatmaps of context frames.
-
-    The input to the CRNN is a set of heatmaps at times t-k, ..., t, ...t+k, one heatmap for each
-    timepoint/keypoint
-
-    The output of the CRNN is a single heatmap for each keypoint
-
-    """
-
-    def __init__(
-        self,
-        num_filters_for_upsampling: int,
-        num_keypoints: int,
-        upsampling_factor: Literal[1, 2] = 2,
-        hkernel: int = 2,
-        hstride: int = 2,
-        hpad: int = 0,
-        nfilters_channel: int = 16,
-    ) -> None:
-        """Upsampling Convolutional RNN - initialize input and hidden weights."""
-
-        super().__init__()
-        self.upsampling_factor = upsampling_factor
-        self.pixel_shuffle = nn.PixelShuffle(2)
-        if self.upsampling_factor == 2:
-            self.W_pre = HeatmapTracker.create_double_upsampling_layer(
-                in_channels=num_filters_for_upsampling // 4,
-                out_channels=num_keypoints,
-            )
-            in_channels_rnn = num_keypoints
-        else:
-            in_channels_rnn = num_filters_for_upsampling // 4
-
-        self.W_f = HeatmapTracker.create_double_upsampling_layer(
-            in_channels=in_channels_rnn,
-            out_channels=num_keypoints,
-        )
-        H_f_layers = []
-        H_f_layers.append(
-            nn.Conv2d(
-                in_channels=num_keypoints,
-                out_channels=num_keypoints * nfilters_channel,
-                kernel_size=(hkernel, hkernel),
-                stride=(hstride, hstride),
-                padding=(hpad, hpad),
-                groups=num_keypoints,
-            )
-        )
-        H_f_layers.append(
-            nn.ConvTranspose2d(
-                in_channels=num_keypoints * nfilters_channel,
-                out_channels=num_keypoints,
-                kernel_size=(hkernel, hkernel),
-                stride=(hstride, hstride),
-                padding=(hpad, hpad),
-                output_padding=(hpad, hpad),
-                groups=num_keypoints,
-            )
-        )
-        self.H_f = nn.Sequential(*H_f_layers)
-
-        self.W_b = HeatmapTracker.create_double_upsampling_layer(
-            in_channels=in_channels_rnn,
-            out_channels=num_keypoints,
-        )
-        H_b_layers = []
-        H_b_layers.append(
-            nn.Conv2d(
-                in_channels=num_keypoints,
-                out_channels=num_keypoints * nfilters_channel,
-                kernel_size=(hkernel, hkernel),
-                stride=(hstride, hstride),
-                padding=(hpad, hpad),
-                groups=num_keypoints,
-            )
-        )
-        H_b_layers.append(
-            nn.ConvTranspose2d(
-                in_channels=num_keypoints * nfilters_channel,
-                out_channels=num_keypoints,
-                kernel_size=(hkernel, hkernel),
-                stride=(hstride, hstride),
-                padding=(hpad, hpad),
-                output_padding=(hpad, hpad),
-                groups=num_keypoints,
-            )
-        )
-        self.H_b = nn.Sequential(*H_b_layers)
-        self.initialize_layers()
-        if self.upsampling_factor == 2:
-            self.layers = torch.nn.ModuleList([self.W_pre, self.W_f, self.H_f, self.W_b, self.H_b])
-        else:
-            self.layers = torch.nn.ModuleList([self.W_f, self.H_f, self.W_b, self.H_b])
-
-    def initialize_layers(self):
-        if self.upsampling_factor == 2:
-            torch.nn.init.xavier_uniform_(self.W_pre.weight, gain=1.0)
-            torch.nn.init.zeros_(self.W_pre.bias)
-
-        torch.nn.init.xavier_uniform_(self.W_f.weight, gain=1.0)
-        torch.nn.init.zeros_(self.W_f.bias)
-        for index, layer in enumerate(self.H_f):
-            torch.nn.init.xavier_uniform_(layer.weight, gain=1.0)
-            torch.nn.init.zeros_(layer.bias)
-
-        torch.nn.init.xavier_uniform_(self.W_b.weight, gain=1.0)
-        torch.nn.init.zeros_(self.W_b.bias)
-        for index, layer in enumerate(self.H_b):
-            torch.nn.init.xavier_uniform_(layer.weight, gain=1.0)
-            torch.nn.init.zeros_(layer.bias)
-
-    def forward(
-        self,
-        representations: TensorType["frames", "batch", "features", "rep_height", "rep_width"]
-    ) -> TensorType["batch", "num_keypoints", "heatmap_height", "heatmap_width"]:
-
-        # expand representations in spatial domain using pixel shuffle to create heatmaps
-        if self.upsampling_factor == 2:
-            # upsample once more before passing through RNN
-            # need to reshape to push through convolution layers
-            frames, batch, features, rep_height, rep_width = representations.shape
-            frames_batch_shape = batch * frames
-            representations_batch_frames: TensorType[
-                "batch*frames", "features", "rep_height", "rep_width"
-            ] = representations.reshape(frames_batch_shape, features, rep_height, rep_width)
-            x_tensor = self.W_pre(self.pixel_shuffle(representations_batch_frames))
-            x_tensor = x_tensor.reshape(
-                frames,
-                batch,
-                x_tensor.shape[1],
-                x_tensor.shape[2],
-                x_tensor.shape[3],
-            )
-        else:
-            x_tensor = self.pixel_shuffle(representations)
-
-        # push heatmaps through CRNN
-        x_f = self.W_f(x_tensor[0])
-        for frame_batch in x_tensor[1:]:  # forward pass
-            x_f = self.W_f(frame_batch) + self.H_f(x_f)
-        x_tensor_b = torch.flip(x_tensor, dims=[0])
-        x_b = self.W_b(x_tensor_b[0])
-        for frame_batch in x_tensor_b[1:]:  # backwards pass
-            x_b = self.W_b(frame_batch) + self.H_b(x_b)
-
-        # average forward/backward heatmaps
-        heatmaps = (x_f + x_b) / 2
-
-        return heatmaps
